Return Idle instead of Off, No preset when in manual mode
hvac_action: Return Idle when system is in not turned Off (e.g. in Heat, Cool) but not actively cooling or heating.  Return off is system is off and no active operation.

preset_mode: When system is in manual mode, there is an active schedule "Manual Mode 0", but since this isn't a preset, return empty string instead.